+++ httpcore/_types.py
@@ -6,6 +6,6 @@ from typing import Dict, List, Optional, Tuple, Union
 
 StrOrBytes = Union[str, bytes]
 Origin = Tuple[bytes, bytes, int]
-URL = Tuple[bytes, bytes, int, bytes]
+URL = Tuple[bytes, bytes, Optional[int], bytes]
 Headers = List[Tuple[bytes, bytes]]
 TimeoutDict = Dict[str, Optional[float]]