+++ tests/async_tests/test_interfaces.py
@@ -186,11 +186,13 @@ async def test_http_proxy(
     method = b"GET"
     url = (b"http", b"example.org", 80, b"/")
     headers = [(b"host", b"example.org")]
-    # Tunnel requires the host header to be present,
-    # Forwarding will use the request headers
-    proxy_headers = headers if proxy_mode == "TUNNEL_ONLY" else None
+    max_connections = 1
+    max_keepalive = 2
     async with httpcore.AsyncHTTPProxy(
-        proxy_server, proxy_headers=proxy_headers, proxy_mode=proxy_mode
+        proxy_server,
+        proxy_mode=proxy_mode,
+        max_connections=max_connections,
+        max_keepalive=max_keepalive,
     ) as http:
         http_version, status_code, reason, headers, stream = await http.request(
             method, url, headers
@@ -205,25 +207,31 @@ async def test_http_proxy(
 # mitmproxy does not support forwarding HTTPS requests
 @pytest.mark.parametrize("proxy_mode", ["DEFAULT", "TUNNEL_ONLY"])
 @pytest.mark.usefixtures("async_environment")
+@pytest.mark.parametrize("http2", [False, True])
 async def test_proxy_https_requests(
     proxy_server: typing.Tuple[bytes, bytes, int],
     ca_ssl_context: ssl.SSLContext,
     proxy_mode: str,
+    http2: bool,
 ) -> None:
     method = b"GET"
     url = (b"https", b"example.org", 443, b"/")
-    headers = proxy_headers = [(b"host", b"example.org")]
+    headers = [(b"host", b"example.org")]
+    max_connections = 1
+    max_keepalive = 2
     async with httpcore.AsyncHTTPProxy(
         proxy_server,
-        proxy_headers=proxy_headers,
         proxy_mode=proxy_mode,
         ssl_context=ca_ssl_context,
+        max_connections=max_connections,
+        max_keepalive=max_keepalive,
+        http2=http2,
     ) as http:
         http_version, status_code, reason, headers, stream = await http.request(
             method, url, headers
         )
         _ = await read_body(stream)
 
-        assert http_version == b"HTTP/1.1"
+        assert http_version == (b"HTTP/2" if http2 else b"HTTP/1.1")
         assert status_code == 200
         assert reason == b"OK"